requirements.txt
Flask
Flask-JWT
Flask-RESTful
Flask-SQLAlchemy
mysqlclient
flasgger
Flask-Excel
pybars3
pdfkit
num2words
flask-cors